+++ Specs/Scene/PointCloud3DTileContentSpec.js
@@ -441,7 +441,8 @@ defineSuite([
         });
     });
 
-    it('attenuates points based on geometric error', function() {
+    var noAttenuationPixelCount = 16;
+    function attenuationTest(postLoadCallback) {
         var scene = createScene({
             canvas : createCanvas(10, 10)
         });
@@ -451,159 +452,103 @@ defineSuite([
         scene.camera.zoomIn(6);
 
         return Cesium3DTilesTester.loadTileset(scene, pointCloudNoColorUrl).then(function(tileset) {
-            var noAttenuationPixelCount = 16;
-
-            tileset.pointShading.attenuation = true;
-            tileset.pointShading.geometricErrorScale = 1.0;
-            tileset.pointShading.maximumAttenuation = undefined;
-            tileset.pointShading.baseResolution = undefined;
-            tileset.maximumScreenSpaceError = 16;
-            expect(scene).toRenderPixelCountGreaterThan(noAttenuationPixelCount);
-
+            tileset.pointCloudShading.eyeDomeLighting = false;
+            postLoadCallback(scene, tileset);
             scene.destroyForSpecs();
+        });
+    }
+
+    it('attenuates points based on geometric error', function() {
+        return attenuationTest(function(scene, tileset) {
+            tileset.pointCloudShading.attenuation = true;
+            tileset.pointCloudShading.geometricErrorScale = 1.0;
+            tileset.pointCloudShading.maximumAttenuation = undefined;
+            tileset.pointCloudShading.baseResolution = undefined;
+            tileset.maximumScreenSpaceError = 16;
+            expect(scene).toRenderPixelCountAndCall(function(pixelCount) {
+                expect(pixelCount).toBeGreaterThan(noAttenuationPixelCount);
+            });
         });
     });
 
     it('modulates attenuation using the tileset screen space error', function() {
-        var scene = createScene({
-            canvas : createCanvas(10, 10)
-        });
-        var center = new Cartesian3.fromRadians(centerLongitude, centerLatitude, 5.0);
-        scene.camera.lookAt(center, new HeadingPitchRange(0.0, -1.57, 5.0));
-        scene.fxaa = false;
-        scene.camera.zoomIn(6);
-
-        return Cesium3DTilesTester.loadTileset(scene, pointCloudNoColorUrl).then(function(tileset) {
-            var noAttenuationPixelCount = 16;
-
-            tileset.pointShading.attenuation = true;
-            tileset.pointShading.geometricErrorScale = 1.0;
-            tileset.pointShading.maximumAttenuation = undefined;
-            tileset.pointShading.baseResolution = undefined;
+        return attenuationTest(function(scene, tileset) {
+            tileset.pointCloudShading.attenuation = true;
+            tileset.pointCloudShading.geometricErrorScale = 1.0;
+            tileset.pointCloudShading.maximumAttenuation = undefined;
+            tileset.pointCloudShading.baseResolution = undefined;
             tileset.maximumScreenSpaceError = 1;
-            expect(scene).toRenderPixelCount(noAttenuationPixelCount);
-
-            scene.destroyForSpecs();
+            expect(scene).toRenderPixelCountAndCall(function(pixelCount) {
+                expect(pixelCount).toEqual(noAttenuationPixelCount);
+            });
         });
     });
 
     it('modulates attenuation using the maximumAttenuation parameter', function() {
-        var scene = createScene({
-            canvas : createCanvas(10, 10)
-        });
-        var center = new Cartesian3.fromRadians(centerLongitude, centerLatitude, 5.0);
-        scene.camera.lookAt(center, new HeadingPitchRange(0.0, -1.57, 5.0));
-        scene.fxaa = false;
-        scene.camera.zoomIn(6);
-
-        return Cesium3DTilesTester.loadTileset(scene, pointCloudNoColorUrl).then(function(tileset) {
-            var noAttenuationPixelCount = 16;
-
-            tileset.pointShading.attenuation = true;
-            tileset.pointShading.geometricErrorScale = 1.0;
-            tileset.pointShading.maximumAttenuation = 1;
-            tileset.pointShading.baseResolution = undefined;
+        return attenuationTest(function(scene, tileset) {
+            tileset.pointCloudShading.attenuation = true;
+            tileset.pointCloudShading.geometricErrorScale = 1.0;
+            tileset.pointCloudShading.maximumAttenuation = 1;
+            tileset.pointCloudShading.baseResolution = undefined;
             tileset.maximumScreenSpaceError = 16;
-            expect(scene).toRenderPixelCount(noAttenuationPixelCount);
-
-            scene.destroyForSpecs();
+            expect(scene).toRenderPixelCountAndCall(function(pixelCount) {
+                expect(pixelCount).toEqual(noAttenuationPixelCount);
+            });
         });
     });
 
     it('modulates attenuation using the baseResolution parameter', function() {
-        var scene = createScene({
-            canvas : createCanvas(10, 10)
-        });
-        var center = new Cartesian3.fromRadians(centerLongitude, centerLatitude, 5.0);
-        scene.camera.lookAt(center, new HeadingPitchRange(0.0, -1.57, 5.0));
-        scene.fxaa = false;
-        scene.camera.zoomIn(6);
-
-        return Cesium3DTilesTester.loadTileset(scene, pointCloudNoColorUrl).then(function(tileset) {
-            var noAttenuationPixelCount = 16;
-
-            tileset.pointShading.attenuation = true;
-            tileset.pointShading.geometricErrorScale = 1.0;
-            tileset.pointShading.maximumAttenuation = undefined;
-            tileset.pointShading.baseResolution = CesiumMath.EPSILON20;
+        return attenuationTest(function(scene, tileset) {
+            tileset.pointCloudShading.attenuation = true;
+            tileset.pointCloudShading.geometricErrorScale = 1.0;
+            tileset.pointCloudShading.maximumAttenuation = undefined;
+            tileset.pointCloudShading.baseResolution = CesiumMath.EPSILON20;
             tileset.maximumScreenSpaceError = 16;
-            expect(scene).toRenderPixelCount(noAttenuationPixelCount);
-
-            scene.destroyForSpecs();
+            expect(scene).toRenderPixelCountAndCall(function(pixelCount) {
+                expect(pixelCount).toEqual(noAttenuationPixelCount);
+            });
         });
     });
 
     it('modulates attenuation using the baseResolution parameter', function() {
-        var scene = createScene({
-            canvas : createCanvas(10, 10)
-        });
-        var center = new Cartesian3.fromRadians(centerLongitude, centerLatitude, 5.0);
-        scene.camera.lookAt(center, new HeadingPitchRange(0.0, -1.57, 5.0));
-        scene.fxaa = false;
-        scene.camera.zoomIn(6);
-
-        return Cesium3DTilesTester.loadTileset(scene, pointCloudNoColorUrl).then(function(tileset) {
-            var noAttenuationPixelCount = 16;
-
+        return attenuationTest(function(scene, tileset) {
             // pointCloudNoColorUrl is a single tile with GeometricError = 0,
             // which results in default baseResolution being computed
-            tileset.pointShading.attenuation = true;
-            tileset.pointShading.geometricErrorScale = 1.0;
-            tileset.pointShading.maximumAttenuation = undefined;
-            tileset.pointShading.baseResolution = CesiumMath.EPSILON20;
+            tileset.pointCloudShading.attenuation = true;
+            tileset.pointCloudShading.geometricErrorScale = 1.0;
+            tileset.pointCloudShading.maximumAttenuation = undefined;
+            tileset.pointCloudShading.baseResolution = CesiumMath.EPSILON20;
             tileset.maximumScreenSpaceError = 16;
-            expect(scene).toRenderPixelCount(noAttenuationPixelCount);
-
-            scene.destroyForSpecs();
+            expect(scene).toRenderPixelCountAndCall(function(pixelCount) {
+                expect(pixelCount).toEqual(noAttenuationPixelCount);
+            });
         });
     });
 
     it('modulates attenuation using the geometricErrorScale parameter', function() {
-        var scene = createScene({
-            canvas : createCanvas(10, 10)
-        });
-        var center = new Cartesian3.fromRadians(centerLongitude, centerLatitude, 5.0);
-        scene.camera.lookAt(center, new HeadingPitchRange(0.0, -1.57, 5.0));
-        scene.fxaa = false;
-        scene.camera.zoomIn(6);
-
-        return Cesium3DTilesTester.loadTileset(scene, pointCloudNoColorUrl).then(function(tileset) {
-            var noAttenuationPixelCount = 16;
-
-            tileset.pointShading.attenuation = true;
-            tileset.pointShading.geometricErrorScale = 0.0;
-            tileset.pointShading.maximumAttenuation = undefined;
-            tileset.pointShading.baseResolution = undefined;
+        return attenuationTest(function(scene, tileset) {
+            tileset.pointCloudShading.attenuation = true;
+            tileset.pointCloudShading.geometricErrorScale = 0.0;
+            tileset.pointCloudShading.maximumAttenuation = undefined;
+            tileset.pointCloudShading.baseResolution = undefined;
             tileset.maximumScreenSpaceError = 1;
-            expect(scene).toRenderPixelCount(noAttenuationPixelCount);
-
-            scene.destroyForSpecs();
+            expect(scene).toRenderPixelCountAndCall(function(pixelCount) {
+                expect(pixelCount).toEqual(noAttenuationPixelCount);
+            });
         });
     });
 
-
     it('attenuates points based on geometric error in 2D', function() {
-        var scene = createScene({
-            canvas : createCanvas(10, 10)
-        });
-        var center = new Cartesian3.fromRadians(centerLongitude, centerLatitude, 5.0);
-        scene.camera.lookAt(center, new HeadingPitchRange(0.0, -1.57, 5.0));
-        scene.fxaa = false;
-        scene.camera.zoomIn(6);
-
-        return Cesium3DTilesTester.loadTileset(scene, pointCloudNoColorUrl).then(function(tileset) {
-
-            var noAttenuationPixelCount = 16;
-
+        return attenuationTest(function(scene, tileset) {
             scene.morphTo2D(0);
-            tileset.pointShading.attenuation = true;
-            tileset.pointShading.geometricErrorScale = 1.0;
-            tileset.pointShading.maximumAttenuation = undefined;
-            tileset.pointShading.baseResolution = undefined;
+            tileset.pointCloudShading.attenuation = true;
+            tileset.pointCloudShading.geometricErrorScale = 1.0;
+            tileset.pointCloudShading.maximumAttenuation = undefined;
+            tileset.pointCloudShading.baseResolution = undefined;
             tileset.maximumScreenSpaceError = 16;
-            expect(scene).toRenderPixelCountGreaterThan(noAttenuationPixelCount);
-
-            scene.destroyForSpecs();
+            expect(scene).toRenderPixelCountAndCall(function(pixelCount) {
+                expect(pixelCount).toBeGreaterThan(noAttenuationPixelCount);
+            });
         });
     });